Corrected variable declaration from byte to int
Corrected variable declaration from byte to int

+++ Github_Tutorial.ino
@@ -21,7 +21,8 @@ void setup()
 
 void loop() 
 {
-  byte myValue = 0;
+//Changed byte to int
+  int myValue = 0;
   myValue = analogRead(A0);
   
   Serial.print("The value is: ");
